requirements.txt
discord.py
google-api-python-client
google-auth
apscheduler
pytz
python-dateutil
python-dotenv
aiosqlite
gspread
pandas
openpyxl
